create WrongAboutOffer interface
